View total spending on the dashboard

+++ src/main/java/com/expense/repository/RevenueMonthRepository.java
@@ -12,7 +12,7 @@ import org.springframework.data.repository.CrudRepository;
 public interface RevenueMonthRepository extends CrudRepository<monthData,Long>{
         
     @Query(value = "SELECT revenue FROM month_data WHERE month=:month and year=:year and user_id=:id", nativeQuery = true)
-    Optional<Double> getRevenue(Integer month,Integer year,Integer id);
+    Optional<Double> getRevenue(Integer month,Integer year,Long id);
 
     @Query(value = "SELECT year year_value,ROUND(sum(md.month_salary)::numeric,2) total_earnings, ROUND(SUM(md.total)::numeric,2) total_bill_year,  ROUND((sum(md.month_salary)-SUM(md.total))::numeric,2) as money_saved FROM month_data md where md.user_id = :id GROUP by md.year", nativeQuery = true)
     List<Map<String,Object>> geyAllYear(Long id);
